Changend signatures for gzip algorithm

+++ scil/code/algo/algo-gzip.c
@@ -18,13 +18,13 @@
 #include <zlib.h>
 
 #pragma GCC diagnostic ignored "-Wunused-parameter"
-int scil_gzip_compress(const scil_context* ctx, byte* restrict dest, uint64_t* restrict dest_size, const byte*restrict source, const size_t source_size){
+int scil_gzip_compress(const scil_context* ctx, byte* restrict dest, size_t* restrict dest_size, const byte*restrict source, const size_t source_size){
 
   return compress( (Bytef*)dest, dest_size, (Bytef*)source, (uLong)(source_size) ) == Z_OK ;
 }
 
 #pragma GCC diagnostic ignored "-Wunused-parameter"
-int scil_gzip_decompress(const scil_context* ctx, byte*restrict dest, uint64_t exp_size, const byte*restrict source, const size_t source_size){
+int scil_gzip_decompress(const scil_context* ctx, byte*restrict dest, size_t exp_size, const byte*restrict source, const size_t source_size){
 
     uLongf dest_isize = exp_size;
     int ret = uncompress( (Bytef*)dest, & dest_isize, (Bytef*)source, (uLong)source_size);
